nvim: setup options and mapping
as far me configures for 'nvim 0.5' I choose 'wincent' as
reference for writing init.lua.

the first step is to configure the options. some differs with wincent, I
added more options.

the second step is to configure the mappings. the maps-keys I take from
my old vimrc.

TODO:of course there is still a lot of improvement that need to be done
in many place.
> configure vim directory structure
> plugins. Is need to use 'packer' or use base vim standard-plugin? as
mentioned at man-page user_05.txt.

link: https://github.com/wincent/wincent