nvim - [keymap]: use best practice writing
I know it's not a big deal, but to get used to writing without having to
abbreviate something is more objective on writing the config.

I'm not only one reading this config or setting I wrote, it's just a
good habits from an initial stage of less complicated features and
configuration.

e.g:

<nop>   >> <Nop>
:hi     >> :higlight

above is a small example that my be a trivial, but it is so annoying to
remember abbreviations that should not be necessary.

TODO:
-- telescope mapping doesn't work if I open a new buffer, maybe it
caused by use 'vim.api.nvim_buf_set_keymap() functions. I'll observe
first.