emacs [config]: complete emacs From Scratch #7 Config Everthing with Org Babel -2
-- Use org headers config

    "#+PROPERTY: header-args:emacs-lisp :tangle ./init-new.el"

Whenever you change your org file (Emacs.org) you need to reevaluate it
so it does get added all of blocks on the file. with 'M-x org-mode'.

- It's better to always to reevaluate the org file and run the tangle
command in every changes org-file.

- It's better to add 'auto-revert-mode' as default in 'init.el' to
Replace current buffer text with the text of the visited file on disk.

-- Add auto tangle configuration
FIXME: It's seems 'add-hook' not run automatically, 'new-init.el' doesn't
change automatically.

-- Configure babel languages
In a moment it just 2 languages as I enable, just for an example.