refactor[emacs]: get rid flycheck warning to free variables
some variables in Emacs are 'buffer-local'; meaning that each buffer is allowed
to have separate value for that variable that overrides the global values.

if variable is buffer-local, then `setq` sets its local value in the current
buffer and `setq-default` set the global default value.

if variable is not buffer-local, then `setq` and `setq-default` do something
different

ref: https://stackoverflow.com/questions/18172728/the-difference-between-setq-and-setq-default-in-emacs-lisp